Move test to its own file for ease of running in isolation
